qtvcp -update core library/status_label to read toolfile
There is code in there to write the file but hasn't been tested.
This is reworked code from gladevcp's tooledit widget.
A full tool editting widget needs to be written to use this new core library.
For the moment we can now get comments from a specfic tool with status_label